dodalam dane do trasy

+++ app/src/main/java/pl/roszkowska/track/database/StepEntity.java
@@ -11,5 +11,5 @@ public class StepEntity {
     double lat;
     double lon;
     long timestamp;
-    double distanceBetweenLastStep;
+    long distanceBetweenLastStep;
 }
